move item form logic into its own component, import that component into LocationDisplay and add button to display form
